Refactored code
The previously hotchpotch code has been rearranged in the respective files.
ui.hpp : header
ui.cpp : general class' definitions
frame.cpp : class frame definition
