+++ test.cpp
@@ -1,0 +1,50 @@
+#include "ui.hpp"
+
+void main()
+{
+	clrscr();
+
+	ui::init();
+	frame f;
+	f.display();
+
+	getch();
+
+	f << ui::top << 't'
+	  << ui::left << 'l'
+	  << ui::bottom << 'b'
+	  << ui::right << 'r';
+
+	f.setcolor(LIGHTBLUE);
+
+	f.display();
+
+	getch();
+
+	f << (ui::top | ui::left) << (char) 201
+	  << (ui::bottom | ui::left) << (char) 200
+	  << (ui::top | ui::right) << (char) 187
+	  << (ui::bottom | ui::right) << (char) 188
+	  << ui::top << (char) 205
+	  << ui::bottom << (char) 205
+	  << ui::left << (char) 186
+	  << ui::right << (char) 186;
+
+	f.setcolor(ui::tcolor);
+
+	f.display();
+
+	getch();
+
+	f.setheight(ui::scr_height/2);
+	getch();
+
+	f.setwidth(ui::scr_width/3);
+	getch();
+
+	f.setcorner_top_left(coord( (ui::scr_width-f.getwidth()) / 2, (ui::scr_height-f.getheight()) / 2));
+	getch();
+
+	f.setvisibility_mode(frame::nosides);
+	getch();
+}
